requirements.txt
PySide6>=6.5
PyYAML>=6.0.1
websocket-client>=1.6.0
